commit UAS

+++ app/Controllers/ProdukController.php
@@ -12,7 +12,7 @@ class ProdukController extends BaseController
 {
     protected $productModel;
     
-    // Tambahkan baris ini agar semua fungsi Form Helper otomatis aktif di semua method/fungsi
+    
     protected $helpers = ['form'];
 
     public function __construct()
@@ -49,35 +49,40 @@ class ProdukController extends BaseController
         return redirect()->to('produk')->with('success', 'Data Berhasil Ditambah');
     }
 
-    public function edit($id)
-    {
-        $dataProduk = $this->productModel->find($id);
+   public function edit($id)
+{
+    $dataProduk = $this->productModel->find($id);
 
-        $dataForm = [
-            'nama' => $this->request->getPost('nama'),
-            'harga' => $this->request->getPost('harga'),
-            'jumlah' => $this->request->getPost('jumlah')
-        ];
+    $dataForm = [
+        'nama'   => $this->request->getPost('nama'),
+        'harga'  => $this->request->getPost('harga'),
+        'jumlah' => $this->request->getPost('jumlah'),
+        'foto'   => $dataProduk['foto'] 
+    ];
 
-        if ($this->request->getPost('check') == 1) {
-            if ($dataProduk['foto'] != '' && file_exists("img/" . $dataProduk['foto'])) {
-                unlink("img/" . $dataProduk['foto']);
-            }
+    $dataFoto = $this->request->getFile('foto');
 
-            $dataFoto = $this->request->getFile('foto');
 
-            if ($dataFoto->isValid()) {
-                $fileName = $dataFoto->getRandomName();
-                $dataFoto->move('img/', $fileName);
-
-                $dataForm['foto'] = $fileName;
-            }
+    if ($this->request->getPost('check') == 1 && $dataFoto && $dataFoto->isValid() && !$dataFoto->hasMoved()) {
+        
+        
+        if ($dataProduk['foto'] != '' && file_exists("img/" . $dataProduk['foto'])) {
+            unlink("img/" . $dataProduk['foto']);
         }
 
-        $this->productModel->update($id, $dataForm);
+    
+        $fileName = $dataFoto->getRandomName();
+        $dataFoto->move('img/', $fileName);
 
-        return redirect()->to('produk')->with('success', 'Data Berhasil Diubah');
+     
+        $dataForm['foto'] = $fileName;
     }
+
+
+    $this->productModel->update($id, $dataForm);
+
+    return redirect()->to('produk')->with('success', 'Data Berhasil Diubah');
+}
 
     public function delete($id)
     {
@@ -90,30 +95,24 @@ class ProdukController extends BaseController
 
     public function download()
     {
-        // Ambil data produk dari database
+       
         $products = $this->productModel->findAll();
 
-        // Render view menjadi HTML
         $html = view('produk/download_pdf', [
             'products' => $products
         ]);
 
-        // Nama file PDF
+     
         $filename = date('Y-m-d-H-i-s') . '-produk.pdf';
-
-        // Inisialisasi Dompdf
+    
         $dompdf = new Dompdf();
-
-        // Load HTML ke Dompdf
+     
         $dompdf->loadHtml($html);
-
-        // Setting ukuran kertas dan orientasi
+     
         $dompdf->setPaper('A4', 'portrait');
-
-        // Generate PDF
+   
         $dompdf->render();
 
-        // Download / tampilkan PDF
         $dompdf->stream($filename, [
             'Attachment' => true
         ]);
